adding a ship class -> enemy and player + remove sisgev on collision + first enemy ship

+++ sources/entities/Projectiles/Projectile.cpp
@@ -22,11 +22,13 @@ void Entities::Projectile::update() {
 }
 
 void Entities::Projectile::onCollision(ENTITY *other) {
+    if (!other->isEnabled)
+        return;
 	if (dynamic_cast<Entities::Asteroid *>(other) != nullptr)
 		this->scope->entityManager.remove(this);
-	else if (dynamic_cast<Entities::PlayerShip *>(other) != nullptr)
-		if (other->id != this->originId) {
-			dynamic_cast<Entities::PlayerShip *>(other)->takeDamage(this->damage);
+	else if (dynamic_cast<Entities::Ship *>(other) != nullptr)
+		if (other->id != this->originId && dynamic_cast<Entities::Ship *>(other)->getTeam() == Entities::Ship::GAME) {
+			dynamic_cast<Entities::Ship *>(other)->takeDamage(this->damage);
 			this->scope->entityManager.remove(this);
 		}
 }
